% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_index.R
\name{collect_testthat_source_info}
\alias{collect_testthat_source_info}
\title{Collect source info on test_that calls}
\usage{
collect_testthat_source_info(infiles)
}
\arguments{
\item{infiles}{a \code{character} vector of file paths}
}
\value{
a \code{data.frame} similar to what \code{\link[=getParseData]{getParseData()}} returns
}
\description{
Collect source info on test_that calls
}
