% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_skeleton.R
\name{test_skeleton}
\alias{test_skeleton}
\title{Create a test skeleton file  for the currently open .R file}
\usage{
test_skeleton(fname, open = TRUE)
}
\arguments{
\item{open}{Should the test file be opend after it is created?}
}
\description{
If the file currently open in the Rstudio editor is called \code{my_function.R},
this creates the file \code{/tests/testthat/test_my_function.R} and fills it
with a basic test skeleton.
}

