% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_coverage.R
\name{extract_test_that_desc}
\alias{extract_test_that_desc}
\title{Extract 'desc' arguments from all test_that functions from .R script files}
\usage{
extract_test_that_desc(infile)
}
\arguments{
\item{infile}{character. Patht to an .R script file, or a list of such paths;
usually created with list.files("/path/to/directory")}
}
\value{
content of the 'desc' arguments of test_that functions
}
\description{
Extract 'desc' arguments from all test_that functions from .R script files
}

