% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_index.R
\name{extract_testthat_parse_data}
\alias{extract_testthat_parse_data}
\title{Extract test_that() call related parse data (see \code{\link[=getParseData]{getParseData()}})}
\usage{
extract_testthat_parse_data(exp)
}
\arguments{
\item{exp}{a \code{character} scalar (to be parsed) or an \code{expression} or a
\code{list} of \code{expressions} (as returned by \code{\link[=parse]{parse()}}.}
}
\value{
a \code{data.frame}. See \code{\link[=getParseData]{getParseData()}}
}
\description{
Extract test_that() call related parse data (see \code{\link[=getParseData]{getParseData()}})
}
