% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_coverage.R
\name{get_all_functions}
\alias{get_all_functions}
\title{Get all functions of a package}
\usage{
get_all_functions(pkg = ".")
}
\arguments{
\item{pgk}{path to the package}
}
\value{
a character vector
}
\description{
Get all functions of a package
}
