% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_coverage.R
\name{get_exported_functions}
\alias{get_exported_functions}
\title{Get exported functions of a package}
\usage{
get_exported_functions(pkg = ".")
}
\arguments{
\item{pkg}{path to package}
}
\value{
a character vector
}
\description{
Lists the functions exported by a package (according to the NAMESPACE file)
}
