% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_functions.R
\name{get_package_functions}
\alias{get_package_functions}
\title{Get functions defined by package}
\usage{
get_package_functions(pgk = ".")
}
\arguments{
\item{pgk}{}
}
\description{
Get functions defined by package
}

