% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_coverage.R
\name{get_test_coverage}
\alias{get_test_coverage}
\title{Get Test Coverage of Package}
\usage{
get_test_coverage(pkg = ".", from_tags = TRUE, from_desc = TRUE)
}
\arguments{
\item{pkg}{path to package}

\item{from_tags}{Checks the files if your test directory for testthis tags.
Speicifically, if you have the comment \code{#* @testing myfunction} in any
of your test files, myfunction will be marked as tested.}

\item{from_desc}{Checks the \code{desc} argument \code{test_that(...)} of
the tests in your test directory for functions names. E.g. if you have a
testfile that contains \code{test_that("myfunction works"), ...}, myfunction
will be marked as tested.}
}
\description{
This extracts the test coverage of the target package (usually the package
you are working on). Bear in mind that testthis uses a checklist-approach
for this, and depends that you either put the function name in your
\code{test_that} calls, or used test_this tags. If you want automatic
analysis of test coverage, you must look in other packages such as
\code{covr}.
}
