% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_index.R
\name{get_test_index}
\alias{get_test_index}
\title{Guess tested functions based on contents of test_that desc argument}
\usage{
get_test_index()
}
\value{
A \code{test_index} \code{data.frame}
}
\description{
Display the \code{desc} arguments of all \code{test_that(} calls in the
\code{tests/testthat} directory of a package`
}
