% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_functions.R
\name{get_tested_functions}
\alias{get_tested_functions}
\title{Get functions defined in package}
\usage{
get_tested_functions(pkg = ".")
}
\arguments{
\item{pkg}{}
}
\description{
Get functions defined in package
}

