% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tests.R
\name{open_tests}
\alias{open_tests}
\title{Open associated test_file}
\usage{
open_tests()
}
\description{
If the currently open file in the Rstudio editor is called \code{myfun.R} this
opens /code{tests/testthat/test_myfun.R} in a new tab.
}
