% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_testdata.R
\name{read_testdata}
\alias{read_testdata}
\alias{find_testdata}
\title{Find / Read Test Data}
\usage{
read_testdata(infile, subdir = NULL)

find_testdata(..., path = ".", must_exist = FALSE)
}
\arguments{
\item{infile}{rds file to read (must end in .rds, otherwise .rds ending is
automatically added)}

\item{subdir}{Character scalar. Subdirectory of \file{testdata} to save
to / read from.}

\item{...}{\R objects to save to the \file{testdata} dir. If empty,
an empty directory is created.}

\item{path}{The start directory}

\item{must_exist}{\code{logical} scalar. Assert that path specified in \code{...}
exists}
}
\value{
\code{read_testdata()} returns a single \R object

\code{find_testdata()} returns the normalized path to a file in a
in the testdata directory
}
\description{
Find / Read Test Data
}
\seealso{
\code{\link[=use_testdata]{use_testdata()}}
}
