% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taglist.R
\name{taglist}
\alias{taglist}
\title{Testthis Taglist}
\usage{
taglist(dat)
}
\arguments{
\item{dat}{Possible tags:
   testfile *path_to_file* ... Put this in a file to manually set the name of an associated
   test file (relevant for \code{test_this}, \code{open_tests}, etc...)

   testing *function_name* ... put this in a testfile to mark the current file includes
   tests for a functiuon}
}
\description{
Testthis Taglist
}
