% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_index.R
\name{test_index}
\alias{test_index}
\title{Guess tested functions based on contents of test_that desc argument}
\usage{
test_index(markers = requireNamespace("rstudioapi", quietly = TRUE))
}
\arguments{
\item{markers}{\code{logical} scalar. Display RStudio markers. If \code{FALSE}, the
index is printed to the console instead.}
}
\value{
A \code{test_index} \code{data.frame} (invisibly if \code{markers == TRUE})
}
\description{
Display the \code{desc} arguments of all \code{test_that(} calls in the
\code{tests/testthat} directory of a package`
}
