% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_this.R
\name{test_this}
\alias{test_this}
\title{Test this file}
\usage{
test_this(...)
}
\arguments{
\item{...}{passed on to \code{\link[testthat:test_file]{testthat::test_file()}}}
}
\value{
\code{NULL} (invisibly)
}
\description{
Runs testthat tests in a single .R file. If the file currently open in the
RStudio editor is called \code{my_function.R}, \code{test_this()} calls
\code{testthat::test_file()} on \file{tests/testthat/test_my_function.R}. If
the filename of the currently open file with starts with \code{test_} it will
call \code{testthat::test_file()} on the current file.
}
\details{
This is useful in cases where you don't want to run all tests in a package
via \code{devtools::test()} (CTRL+SHIFT+T).
}
