% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_with_skip.R
\name{test_with_skip}
\alias{test_with_skip}
\title{Execute all test_that tests in a package, except some.}
\usage{
test_with_skip(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[testthat:test_dir]{testthat::test_dir()}} and
\code{\link[covr:package_coverage]{covr::package_coverage()}}}
}
\description{
This is a wrapper for \code{\link[devtools:test]{devtools::test()}} that skips \code{test_*.R}
files that contain the testthis tag \code{#' @skip}.
}
\details{
See the devtools documentation for further info or
\code{vignette("testthis")} for infos on testthis tags.
}
