% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_testdata.R
\name{use_testdata_raw}
\alias{use_testdata_raw}
\title{Create \code{testdata-raw} folder.}
\usage{
use_testdata_raw()
}
\description{
A folder to put scripts in that produce the files in \file{testdata}
}
\seealso{
Other infrastructure: \code{\link{use_test_subdir}},
  \code{\link{use_testdata}}, \code{\link{use_tester}},
  \code{\link{use_testignore}}
}
\concept{infrastructure}
