% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_testignore.R
\name{use_testignore}
\alias{use_testignore}
\title{Tell testthis that a function does not need to be tested}
\usage{
use_testignore(x)
}
\arguments{
\item{x}{a character vector of ignores (function names).}
}
\value{
\code{TRUE} (invisibly).
}
\description{
Tell testthis that a function does not need to be tested
(for \code{\link[=get_test_coverage]{get_test_coverage()}})
}
\examples{

\dontrun{
use_testignore("helperfunction")
}

}
\seealso{
Other infrastructure: \code{\link{use_test_subdir}},
  \code{\link{use_testdata_raw}},
  \code{\link{use_testdata}}, \code{\link{use_tester}}
}
\concept{infrastructure}
