% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_coverage.R
\name{get_test_coverage}
\alias{get_test_coverage}
\title{Get Test Coverage}
\usage{
get_test_coverage(from_tags = TRUE, from_desc = TRUE)
}
\arguments{
\item{from_tags, from_desc}{see \code{\link[=test_coverage]{test_coverage()}}}
}
\description{
Deprecated in favor of \code{\link[=test_coverage]{test_coverage()}}
}
