% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_coverage.R
\name{get_tested_functions}
\alias{get_tested_functions}
\title{Get tested functions of a package}
\usage{
get_tested_functions(pkg, from_tags, from_desc)
}
\arguments{
\item{pkg}{path to package}

\item{from_tags}{Checks the files if your test directory for testthis tags.
Speicifically, if you have the comment \code{#* @testing myfunction} in any
of your test files, myfunction will be marked as tested.}

\item{from_desc}{Checks the \code{desc} argument \code{test_that(...)} of
the tests in your test directory for functions names. E.g. if you have a
testfile that contains \code{test_that("myfunction works"), ...}, myfunction
will be marked as tested.}
}
\value{
a character vector
}
\description{
Get tested functions of a package
}

