% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_index.R
\name{print.test_index}
\alias{print.test_index}
\title{Print Test Index objects}
\usage{
\method{print}{test_index}(x)
}
\arguments{
\item{x}{see \code{\link[=get_test_index]{get_test_index()}}}
}
\value{
\code{x} (invisibly)
}
\description{
Print Test Index objects
}
