% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthis.R
\name{test_this}
\alias{lest_this}
\alias{test_this}
\title{Test this file}
\usage{
test_this()

lest_this()
}
\description{
If the current file is called "function.R", test_this calls
testthat::test_file on "tests/testhat/test_function.R".

If the current file is called "function.R", test_this calls
testthat::test_file on "tests/testhat/test_function.R".
}

