% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_with_skip}
\alias{test_with_skip}
\title{Execute all \pkg{test_that} tests in a package.}
\usage{
test_with_skip(pkg = ".", ...)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information}

\item{...}{additional arguments passed to \code{\link[testthat]{test_dir}}}
}
\description{
This is a wrapper for \code{\link[devtools]{test}} that skips \code{test_*.R}
files that contain the testthis tag "\code{#' @skip}".
}
\details{
See the devtools documentation for further info or
\code{vignette("testthis")} for infos on testthis tags.
}
