% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_test_subdir.R
\name{use_tester}
\alias{use_tester}
\title{Use a tester function}
\usage{
use_tester(path, ignore = TRUE, tester_path = file.path("R",
  "testthis-testers.R"))
}
\arguments{
\item{path}{Name of the subdirectory oft \file{tests/testthat/} for which
to create a tester function.}

\item{ignore}{Logical. Add \code{tester_path} to \file{.Rbuildignore}?}

\item{tester_path}{\R script file in which to store the tester functions}
}
\value{
\code{TRUE} on success (invisibly).
}
\description{
Quickly create an \R script that contains a function for running all tests
in a predefined directory. This function powers the \code{make_tester} option
of \code{\link[=use_test_subdir]{use_test_subdir()}} and you will likely not need to run it manually.
}
\seealso{
Other infrastructure: \code{\link{use_test_subdir}},
  \code{\link{use_testdata_raw}},
  \code{\link{use_testdata}}, \code{\link{use_testignore}}
}
\concept{infrastructure}
